+++ package.json
@@ -2,7 +2,7 @@
   "name": "rbac2",
   "version": "1.1.0",
   "description": "Simple RBAC checker with support for context checks.",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "gulp"
   },
